+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.4.4]
 ### Added
 - Add eventsAttending to the User model and Person type
+- Add migration to create user_affiliations table, model for Affiliation, and mutation resolvers to createAffiliation and deleteAffiliation
 
 ## [1.4.3]
 ### Added